clx/clx.lisp: implement backing-pixmap store
* projection/abstract-projection.lisp: (call-with-projection-context),
if the context has already bound *context-view* then treat it as the
"backing-store" to which we have to draw, and retain the binding,
but call the CONTEXT-UPDATE-FOR-VIEW  protocol method with the actual
window system view.

* clx/clx.lisp: (clx-context): (backing-pixmap): new slot.
(context-close-view): free up backing-pixmap.
(call-in-clx-context): arrange for abstract-projection to draw
on our backing-pixmap instead of the X window.
(maybe-reinitialize-backing-pixmap): new function. creates or
reinitializes the backing pixmap.
(context-make-view, context-update-for-view): used here.
(maybe-paint-from-backing-pixmap): new function.
(content-draw-contents): implement protocol method of the same name.
maybe-paint-from-backing-pixmap used here.