implement backing-pixmap store
this feature is conditionalized on *use-backing-pixmap*, set to NIL by
default. this is only necessary if the X server doesn't support
backingstore for windows. (enabled in a subsequent commit)

* projection/abstract-projection.lisp: (call-with-projection-context),
if the context has already bound *context-view* then treat it as the
"backing-store" to which we have to draw, and retain the binding,
but call the CONTEXT-UPDATE-FOR-VIEW  protocol method with the actual
window system view.

* clx/clx.lisp: (clx-context): (backing-pixmap): new slot.
(context-close-view): free up backing-pixmap.
(call-in-clx-context): arrange for abstract-projection to draw
on our backing-pixmap instead of the X window.
(maybe-reinitialize-backing-pixmap): new function. creates or
reinitializes the backing pixmap.
(context-make-view, context-update-for-view): used here.
(maybe-paint-from-backing-pixmap): new function.
(content-draw-contents): implement protocol method of the same name.
maybe-paint-from-backing-pixmap used here.

[250802]
* clx/clx.lisp: (context-clear-view): also clear backing-pixmap.
(context-fill-view): draw on backing-pixmap instead of window.

[251219]
* clx/clx.lisp: (call-in-clx-context): bind *context-view* if it is
not already bound, check it is the backing-pixmap if bound.

[251221]
* tests/projection/clx.lisp: (:og.clx.ffi): update to use backing
store.

[251231]
clx/clx.lisp: (clx-poly-port): use *context-view* which is bound for
us instead of accessing (context-view),