clx/clx.lisp: (clx-arc-port): try to fix.
needs review

* clx/clx.lisp: (clx-arc-port): rewrite to fix incorrect
usage of the CLX xlib:draw-arc api (it uses radians-extent
instead of the end angle). (radians-mod-pi): to work in
the [-pi, pi] range. this may be improved

[note to self:] the arcs and circles tests don't fail because the radius
is not getting transformed correctly, my BOE can't show (.0 .0 .5) blue
on a black background

[260101]
ensure width is an integer by calling round.  can't call
%round here because radius is declared a number.